feat: add responsive design
