+++ src/pages/not-found-page/not-found.module.css
@@ -1,8 +1,6 @@
 .errorMessageContainer {
   display: flex;
-  align-items: flex-start;
-  flex-direction: column;
-  justify-content: center;
+  align-items: center;
   height: 130px;
   width: 100%;
   background: var(--clr-bg-2);
@@ -33,16 +31,38 @@
 }
 
 .imgNotFound {
-  width: 50%;
+  width: 60%;
 }
 
 .textImage {
+  text-align: center;
   font-weight: 700;
   font-size: 24px;
+  padding: 0 10px;
 }
 
 .backButton {
   font-size: 16px;
   font-weight: 400;
   padding: 15px 30px;
+}
+
+@media (width >=360px) and (width <=800px) {
+  .mainContainer {
+    margin: 20px;
+  }
+
+  .errorTitle {
+    font-size: 20px;
+  }
+
+  .textImage {
+    font-size: 16px;
+  }
+
+  .backButton {
+    font-size: 12px;
+    padding: 8px 15px;
+  }
+
 }
